Unify HttpException into a single shared type.
